+++ CHANGES.md
@@ -5,6 +5,7 @@
 ## Changed
 
 * switch from `fastapi` to `fastapi-slim` to avoid installing unwanted dependencies
+* replace Enum with `Literal` for `FilterLang`
 
 ## [3.0.0a0] - 2024-05-06
 
@@ -13,7 +14,7 @@
 * Add enhanced middleware configuration to the StacApi class, enabling specific middleware options and dynamic addition post-application initialization. ([#442](https://github.com/stac-utils/stac-fastapi/pull/442))
 * Add Response Model to OpenAPI, even if model validation is turned off ([#625](https://github.com/stac-utils/stac-fastapi/pull/625))
 
-## Changes
+## Changed
 
 * Update to pydantic v2 and stac_pydantic v3 ([#625](https://github.com/stac-utils/stac-fastapi/pull/625))
 * Removed internal Search and Operator Types in favor of stac_pydantic Types ([#625](https://github.com/stac-utils/stac-fastapi/pull/625))